+++ ReleaseNotes.md
@@ -25,3 +25,4 @@
 - Add MileGetWindowsVersion function.
 - Add MileLoadLibraryFromSystem32 function.
 - Add Mile::WinRT::EmptyValue template function.
+- Add Mile::WinRT::Property template class.